+++ client/src/test/java/io/pravega/client/stream/impl/EventStreamWriterTest.java
@@ -85,7 +85,7 @@ public class EventStreamWriterTest extends ThreadPooledTestSuite {
     }
     
     private CompletableFuture<StreamSegmentsWithPredecessors> getReplacement(Segment old, Segment repacement) {
-        Map<SegmentWithRange, List<Integer>> segments = new HashMap<>();
+        Map<SegmentWithRange, List<Long>> segments = new HashMap<>();
         segments.put(new SegmentWithRange(repacement, 0, 1), Collections.singletonList(old.getSegmentNumber()));
         return CompletableFuture.completedFuture(new StreamSegmentsWithPredecessors(segments, ""));
     }
