+++ client/src/test/java/io/pravega/client/stream/impl/ModelHelperTest.java
@@ -37,7 +37,7 @@ public class ModelHelperTest {
     @Rule
     public Timeout globalTimeout = new Timeout(10, TimeUnit.SECONDS);
 
-    private static Segment createSegmentId(String streamName, int number) {
+    private static Segment createSegmentId(String streamName, long number) {
         return new Segment("scope", streamName, number);
     }
 
@@ -70,10 +70,10 @@ public class ModelHelperTest {
 
     @Test
     public void encodeSegmentId() {
-        Segment segment = ModelHelper.encode(ModelHelper.decode(createSegmentId("stream1", 2)));
+        Segment segment = ModelHelper.encode(ModelHelper.decode(createSegmentId("stream1", 2L)));
         assertEquals("stream1", segment.getStreamName());
         assertEquals("scope", segment.getScope());
-        assertEquals(2, segment.getSegmentNumber());
+        assertEquals(2L, segment.getSegmentNumber());
     }
 
     @Test(expected = NullPointerException.class)
@@ -184,8 +184,8 @@ public class ModelHelperTest {
     public void createSuccessorResponse() {
         Controller.SegmentRange segmentRange = createSegmentRange(0.1, 0.5);
 
-        Map<Controller.SegmentRange, List<Integer>> inputMap = new HashMap<>(1);
-        inputMap.put(segmentRange, Arrays.asList(1));
+        Map<Controller.SegmentRange, List<Long>> inputMap = new HashMap<>(1);
+        inputMap.put(segmentRange, Arrays.asList(1L));
 
         Controller.SuccessorResponse successorResponse = ModelHelper.createSuccessorResponse(inputMap).build();
         Assert.assertEquals(1, successorResponse.getSegmentsCount());
